requirements.txt
Flask==2.0.1
werkzeug<3.0.0
mysql-connector-python
